The rasterize method has been moved to TiledRasterRDD
