+++ geopyspark/geotrellis/rdd.py
@@ -95,58 +95,6 @@ class RasterRDD(object):
 
         return cls(geopysc, rdd_type, srdd)
 
-    @classmethod
-    def rasterize(cls, geopysc, rdd_type, geometry, extent, crs, cols, rows,
-                  fill_value, instant=None):
-        """Creates a RasterRDD from a shapely geomety.
-
-        Args:
-            geopysc (GeoPyContext): The GeoPyContext being used this session.
-            rdd_type (str): What the spatial type of the geotiffs are. This is
-                represented by the constants: `SPATIAL` and `SPACETIME`.
-            geometry (str, Polygon): The value to be turned into a raster. Can either be a
-                string or a ``Polygon``. If the value is a string, it must be the WKT string,
-                geometry format.
-            extent (:ref:`extent`): The ``extent`` of the new raster.
-            crs (str): The CRS the new raster should be in.
-            cols (int): The number of cols the new raster should have.
-            rows (int): The number of rows the new raster should have.
-            fill_value (int): The value to fill the raster with.
-
-                Note:
-                    Only the area the raster intersects with the ``extent`` will have this value.
-                    Any other area will be filled with GeoTrellis' NoData value for ``int`` which
-                    is represented in GeoPySpark as the constant, ``NODATAINT``.
-            instant(int, optional): Optional if the data has no time component (ie is ``SPATIAL``).
-                Otherwise, it is requires and represents the time stamp of the data.
-
-        Returns:
-            :class:`~geopyspark.geotrellis.rdd.RasterRDD`
-
-        Raises:
-            TypeError: If ``geometry`` is not a ``str`` or a ``Polygon``; or if there was a
-                mistach in inputs. Mainly, setting the ``rdd_type`` as ``SPATIAL`` but also setting
-                ``instant``.
-        """
-
-        if isinstance(geometry, str):
-            pass
-        elif isinstance(geometry, Polygon):
-            geometry = dumps(geometry)
-        else:
-            raise TypeError(geometry, "Needs to be either a Polygon or a string")
-
-        if instant and rdd_type != SPATIAL:
-            srdd = geopysc._jvm.geopyspark.geotrellis.TemporalRasterRDD.rasterize(
-                geopysc.sc, geometry, extent, crs, instant, cols, rows, fill_value)
-        elif not instant and rdd_type == SPATIAL:
-            srdd = geopysc._jvm.geopyspark.geotrellis.ProjectedRasterRDD.rasterize(
-                geopysc.sc, geometry, extent, crs, cols, rows, fill_value)
-        else:
-            raise TypeError("Abiguous inputs. Given ", instant, " but rdd_type is SPATIAL")
-
-        return cls(geopysc, rdd_type, srdd)
-
     def to_numpy_rdd(self):
         """Converts a RasterRDD to a numpy RDD.
 
@@ -339,6 +287,58 @@ class TiledRasterRDD(object):
                         reserialized_rdd._jrdd, schema, json.dumps(metadata))
 
         return cls(geopysc, key, srdd)
+
+    @classmethod
+    def rasterize(cls, geopysc, rdd_type, geometry, extent, crs, cols, rows,
+                  fill_value, instant=None):
+        """Creates a TiledRasterRDD from a shapely geomety.
+
+        Args:
+            geopysc (GeoPyContext): The GeoPyContext being used this session.
+            rdd_type (str): What the spatial type of the geotiffs are. This is
+                represented by the constants: `SPATIAL` and `SPACETIME`.
+            geometry (str, Polygon): The value to be turned into a raster. Can either be a
+                string or a ``Polygon``. If the value is a string, it must be the WKT string,
+                geometry format.
+            extent (:ref:`extent`): The ``extent`` of the new raster.
+            crs (str): The CRS the new raster should be in.
+            cols (int): The number of cols the new raster should have.
+            rows (int): The number of rows the new raster should have.
+            fill_value (int): The value to fill the raster with.
+
+                Note:
+                    Only the area the raster intersects with the ``extent`` will have this value.
+                    Any other area will be filled with GeoTrellis' NoData value for ``int`` which
+                    is represented in GeoPySpark as the constant, ``NODATAINT``.
+            instant(int, optional): Optional if the data has no time component (ie is ``SPATIAL``).
+                Otherwise, it is requires and represents the time stamp of the data.
+
+        Returns:
+            :class:`~geopyspark.geotrellis.rdd.TiledRasterRDD`
+
+        Raises:
+            TypeError: If ``geometry`` is not a ``str`` or a ``Polygon``; or if there was a
+                mistach in inputs. Mainly, setting the ``rdd_type`` as ``SPATIAL`` but also setting
+                ``instant``.
+        """
+
+        if isinstance(geometry, str):
+            pass
+        elif isinstance(geometry, Polygon):
+            geometry = dumps(geometry)
+        else:
+            raise TypeError(geometry, "Needs to be either a Polygon or a string")
+
+        if instant and rdd_type != SPATIAL:
+            srdd = geopysc._jvm.geopyspark.geotrellis.TemporalTiledRasterRDD.rasterize(
+                geopysc.sc, geometry, extent, crs, instant, cols, rows, fill_value)
+        elif not instant and rdd_type == SPATIAL:
+            srdd = geopysc._jvm.geopyspark.geotrellis.SpatialTiledRasterRDD.rasterize(
+                geopysc.sc, geometry, extent, crs, cols, rows, fill_value)
+        else:
+            raise TypeError("Abiguous inputs. Given ", instant, " but rdd_type is SPATIAL")
+
+        return cls(geopysc, rdd_type, srdd)
 
     def to_numpy_rdd(self):
         """Converts a TiledRasterRDD to a numpy RDD.
